# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/config/yaml": {
            "get": {
                "description": "返回整个 YAML 配置文件内容",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "配置管理"
                ],
                "summary": "获取当前 YAML 配置",
                "responses": {
                    "200": {
                        "description": "返回 YAML 格式的配置内容",
                        "schema": {
                            "$ref": "#/definitions/response.Resp"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.GetConfResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "description": "接收 JSON 格式的 YAML 配置内容并覆盖当前配置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "配置管理"
                ],
                "summary": "更新 YAML 配置",
                "parameters": [
                    {
                        "description": "更新配置",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.PUTConfReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "$ref": "#/definitions/response.Resp"
                        }
                    },
                    "400": {
                        "description": "请求格式错误",
                        "schema": {
                            "$ref": "#/definitions/response.Resp"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/response.Resp"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "request.PUTConfReq": {
            "type": "object",
            "properties": {
                "conf": {
                    "description": "\"yaml配置\"",
                    "type": "string"
                }
            }
        },
        "response.GetConfResp": {
            "type": "object",
            "properties": {
                "conf": {
                    "type": "string"
                }
            }
        },
        "response.Resp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        }
    }
}